fixes and content changes

+++ resources/views/flush.blade.php
@@ -1,10 +1,11 @@
 @include('framework.functions')
 @php
 
-$userID = $_SESSION['vets-user'];
-
-deleteData($userID);
-
+if (!empty($_SESSION['vets-user'])) {
+    if ($userID = $_SESSION['vets-user']) {
+        deleteData($userID);
+    }
+}
 
 header("Location: /");
 die();
